requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-ratelimit==4.1.0
celery==5.3.1
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
pytz==2023.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
